app/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
pandas
# OpenTelemetry packages
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-trace